Migrated metadata caching logic from metadata/__init__.py into metadata/caching.py.

+++ music21/metadata/__init__.py
@@ -7,7 +7,8 @@
 #               Michael Scott Cuthbert
 #
 # Copyright:    Copyright © 2010, 2012 Michael Scott Cuthbert and the music21
-# Project License:      LGPL, see license.txt
+#               Project 
+# License:      LGPL, see license.txt
 #-------------------------------------------------------------------------------
 '''Classes and functions for creating and processing metadata associated with
 scores, works, and fragments, such as titles, movements, authors, publishers,
@@ -39,12 +40,9 @@ The following example creates a :class:`~music21.stream.Stream` object, adds a
 '''
 
 import datetime
-import multiprocessing
 import os
-import pickle
 import re
 import time
-import traceback
 import unittest
 
 from music21 import base
@@ -52,13 +50,16 @@ from music21 import common
 from music21 import exceptions21
 from music21 import freezeThaw
 from music21 import text
+
+from music21.metadata.caching import *
 from music21.metadata.primitives import *
 
-from music21 import environment
-_MOD = "metadata.py"
-environLocal = environment.Environment(_MOD)
 
 #-------------------------------------------------------------------------------
+
+
+from music21 import environment
+environLocal = environment.Environment(os.path.basename(__file__))
 
 
 #-------------------------------------------------------------------------------
@@ -1084,6 +1085,7 @@ class MetadataBundle(object):
             >>> metadataBundle = metadata.MetadataBundle()
             >>> metadataBundle.addFromPaths(
             ...     corpus.getWorkList('bwv66.6'),
+            ...     useCorpus=True,
             ...     useMultiprocessing=False,
             ...     )
             []
@@ -1139,7 +1141,14 @@ class MetadataBundle(object):
             jobProcessor = JobProcessor.process_parallel
         else:
             jobProcessor = JobProcessor.process_serial
-        for results, errors in jobProcessor(jobs):
+        for results, errors, lastJobFilePath, remainingJobs in jobProcessor(
+            jobs):
+            JobProcessor.report(
+                len(jobs),
+                remainingJobs,
+                lastJobFilePath,
+                len(accumulatedErrors),
+                )
             currentIteration += 1
             accumulatedResults.extend(results)
             accumulatedErrors.extend(errors)
@@ -1191,7 +1200,8 @@ class MetadataBundle(object):
     
     def delete(self):
         if self.filePath is not None:
-            os.remove(self.filePath)
+            if os.path.exists(self.filePath):
+                os.remove(self.filePath)
         return self
 
     @classmethod
@@ -1242,6 +1252,7 @@ class MetadataBundle(object):
             >>> metadataBundle = metadata.MetadataBundle()
             >>> metadataBundle.addFromPaths(
             ...     workList,
+            ...     useCorpus=True,
             ...     useMultiprocessing=False,
             ...     )
             []
@@ -1320,6 +1331,7 @@ class MetadataBundle(object):
             >>> mb = metadata.MetadataBundle()
             >>> mb.addFromPaths(
             ...     corpus.getWorkList('bwv66.6'),
+            ...     useCorpus=True,
             ...     useMultiprocessing=False,
             ...     )
             []
@@ -1413,318 +1425,6 @@ class MetadataBundle(object):
             jsf = freezeThaw.JSONFreezer(self)
             return jsf.jsonWrite(filePath)
         return self
-
-
-#-------------------------------------------------------------------------------
-
-
-class MetadataCacheException(exceptions21.Music21Exception):
-    pass
-
-
-def cacheMetadata(domains=('local', 'core', 'virtual')): 
-    '''
-    The core cache is all locally-stored corpus files. 
-    '''
-    from music21 import corpus, metadata
-
-    if not common.isListLike(domains):
-        domains = (domains,)
-
-    timer = common.Timer()
-    timer.start()
-
-    # store list of file paths that caused an error
-    failingFilePaths = []
-
-    domainGetPathsProcedures = {
-       'core': corpus.getCorePaths,
-       'local': corpus.getLocalPaths,
-       'virtual': corpus.getVirtualPaths,
-       }
-
-    # the core cache is based on local files stored in music21
-    # virtual is on-line
-    for domain in domains:
-        # the domain passed here becomes the name of the bundle
-        # determines the file name of the json bundle
-        metadataBundle = metadata.MetadataBundle(domain)
-        if domain not in domainGetPathsProcedures:
-            raise MetadataCacheException('invalid domain provided: {0}'.format(
-                domain))
-        if os.path.exists(metadataBundle.filePath):
-            metadataBundle.read()
-        paths = domainGetPathsProcedures[domain]()
-        environLocal.warn(
-            'metadata cache: starting processing of paths: {0}'.format(
-                len(paths)))
-        # returns any paths that failed to load
-        failingFilePaths += metadataBundle.addFromPaths(
-            paths, printDebugAfter=1) 
-        environLocal.warn(
-            'cache: writing time: {0} md items: {1}'.format(
-                timer, len(metadataBundle.storage)))
-        del metadataBundle
-
-    environLocal.warn('cache: final writing time: {0} seconds'.format(timer))
-    for failingFilePath in failingFilePaths:
-        environLocal.warn('path failed to parse: {0}'.format(failingFilePath))
-
-
-class MetadataCachingJob(object):
-    '''
-    Parses one corpus path, and attempts to extract metadata from it:
-
-    ::
-
-        >>> job = metadata.MetadataCachingJob(
-        ...     'bach/bwv66.6',
-        ...     useCorpus=True,
-        ...     )
-        >>> job()
-        ((<music21.metadata.MetadataEntry: bach_bwv66_6>,), ())
-        >>> results = job.getResults()
-
-    '''
-    
-    ### INITIALIZER ###
-
-    def __init__(self, filePath, jobNumber=0, useCorpus=True):
-        self.filePath = filePath
-        self.filePathErrors = []
-        self.jobNumber = int(jobNumber)
-        self.results = []
-        self.useCorpus = bool(useCorpus)
-
-    ### SPECIAL METHODS ###
-
-    def __call__(self):
-        import gc
-        self.results = []
-        parsedObject = self._parseFilePath()
-        if parsedObject is not None:
-            if 'Opus' in parsedObject.classes:
-                self._parseOpus(parsedObject)
-            else:
-                self._parseNonOpus(parsedObject)
-        del parsedObject
-        gc.collect()
-        return self.getResults(), self.getErrors()
-
-    ### PRIVATE METHODS ###
-
-    def _parseFilePath(self):
-        from music21 import converter
-        from music21 import corpus
-        parsedObject = None
-        try:
-            if self.useCorpus is False:
-                parsedObject = converter.parse(
-                    self.filePath, forceSource=True)
-            else:
-                parsedObject = corpus.parse(
-                    self.filePath, forceSource=True)
-        except Exception, e:
-            environLocal.warn('parse failed: {0}, {1}'.format(
-                self.filePath, str(e)))
-            self.filePathErrors.append(self.filePath)
-        return parsedObject
-
-    def _parseNonOpus(self, parsedObject):
-        try:
-            corpusPath = MetadataBundle.corpusPathToKey(self.filePath)
-            metadata = parsedObject.metadata
-            if metadata is not None:
-                richMetadata = RichMetadata()
-                richMetadata.merge(metadata)
-                richMetadata.update(parsedObject) # update based on Stream
-                environLocal.printDebug(
-                    'updateMetadataCache: storing: {0}'.format(corpusPath))
-                metadataEntry = MetadataEntry(
-                    cacheTime=time.time(),
-                    filePath=self.filePath,
-                    richMetadata=richMetadata,
-                    )
-            else:
-                environLocal.warn(
-                    'addFromPaths: got stream without metadata, '
-                    'creating stub: {0}'.format(
-                        os.path.relpath(self.filePath)))
-                metadataEntry = MetadataEntry(
-                    cacheTime=time.time(),
-                    filePath=self.filePath,
-                    richMetadata=None,
-                    )
-            self.results.append(metadataEntry)
-        except Exception:
-            environLocal.warn('Had a problem with extracting metadata '
-            'for {0}, piece ignored'.format(self.filePath))
-            traceback.print_exc()
-
-    def _parseOpus(self, parsedObject):
-        # need to get scores from each opus?
-        # problem here is that each sub-work has metadata, but there
-        # is only a single source file
-        try:
-            for scoreNumber, score in enumerate(parsedObject.scores):
-                self._parseOpusScore(score, scoreNumber)
-                del score # for memory conservation
-        except Exception as exception:
-            environLocal.warn(
-                'Had a problem with extracting metadata for score {0} '
-                'in {1}, whole opus ignored: {2}'.format(
-                    scoreNumber, self.filePath, str(exception)))
-            traceback.print_exc()
-        else:
-            # Create a dummy metadata entry, representing the entire opus.
-            # This lets the metadata bundle know it has already processed this
-            # entire opus on the next cache update.
-            metadataEntry = MetadataEntry(
-                cacheTime=time.time(),
-                filePath=self.filePath,
-                richMetadata=None,
-                )
-            self.results.append(metadataEntry)
-
-    def _parseOpusScore(self, score, scoreNumber):
-        try:
-            metadata = score.metadata
-            # updgrade metadata to richMetadata
-            richMetadata = RichMetadata()
-            richMetadata.merge(metadata)
-            richMetadata.update(score) # update based on Stream
-            if metadata is None or metadata.number is None:
-                environLocal.warn(
-                    'addFromPaths: got Opus that contains '
-                    'Streams that do not have work numbers: '
-                    '{0}'.format(self.filePath))
-            else:
-                # update path to include work number
-                corpusPath = MetadataBundle.corpusPathToKey(
-                    self.filePath, 
-                    number=metadata.number,
-                    )
-                environLocal.printDebug(
-                    'addFromPaths: storing: {0}'.format(
-                        corpusPath))
-                metadataEntry = MetadataEntry(
-                    cacheTime=time.time(),
-                    filePath=self.filePath,
-                    number=scoreNumber,
-                    richMetadata=richMetadata,
-                    )
-                self.results.append(metadataEntry)
-        except Exception as exception:
-            environLocal.warn(
-                'Had a problem with extracting metadata '
-                'for score {0} in {1}, whole opus ignored: '
-                '{2}'.format(
-                    scoreNumber, self.filePath, str(exception)))
-            traceback.print_exc()
-
-    ### PUBLIC METHODS ###
-
-    def getErrors(self):
-        return tuple(self.filePathErrors)
-
-    def getResults(self):
-        return tuple(self.results)
-
-
-class JobProcessor(object):
-
-    ### PUBLIC METHODS ###
-
-    @staticmethod
-    def process_parallel(jobs, processCount=None):
-        processCount = processCount or multiprocessing.cpu_count() - 1
-        if processCount < 1:
-            processCount = 1
-        remainingJobs = len(jobs)
-        results = []
-        filePathErrors = []
-        job_queue = multiprocessing.JoinableQueue()
-        result_queue = multiprocessing.Queue()
-        workers = [WorkerProcess(job_queue, result_queue) 
-            for i in range(processCount)]
-        for worker in workers:
-            worker.start()
-        if jobs:
-            for job in jobs:
-                job_queue.put(pickle.dumps(job, protocol=0))
-            for i in xrange(len(jobs)):
-                job = pickle.loads(result_queue.get())
-                results = job.getResults()
-                errors = job.getErrors()
-                remainingJobs -= 1
-                JobProcessor.report(
-                    len(jobs),
-                    remainingJobs,
-                    job.filePath,
-                    len(filePathErrors),
-                    )
-                yield results, errors
-        for worker in workers:
-            job_queue.put(None)
-        job_queue.join()
-        result_queue.close()
-        job_queue.close()
-        for worker in workers:
-            worker.join()
-        raise StopIteration
-
-    @staticmethod
-    def process_serial(jobs):
-        remainingJobs = len(jobs)
-        results = []
-        filePathErrors = []
-        for i, job in enumerate(jobs):
-            results, errors = job()
-            remainingJobs -= 1
-            JobProcessor.report(
-                len(jobs),
-                remainingJobs,
-                job.filePath,
-                len(filePathErrors),
-                )
-            yield results, errors
-        raise StopIteration
-
-    @staticmethod
-    def report(totalJobs, remainingJobs, filePath, filePathErrorCount):
-        message = 'updated {0} of {1} files; ' \
-            'total errors: {2} ... last file: {3}'.format(
-                totalJobs - remainingJobs,
-                totalJobs,
-                filePathErrorCount,
-                os.path.relpath(filePath),
-                )
-        environLocal.warn(message)
-
-
-class WorkerProcess(multiprocessing.Process):
-    
-    ### INITIALIZER ###
-
-    def __init__(self, job_queue, result_queue):
-        multiprocessing.Process.__init__(self)
-        self.job_queue = job_queue
-        self.result_queue = result_queue
-        
-    ### PUBLIC METHODS ###
-
-    def run(self):
-        while True:
-            job = self.job_queue.get()
-            # "Poison Pill" causes worker shutdown:
-            if job is None:
-                self.job_queue.task_done()
-                break
-            job = pickle.loads(job)
-            job()
-            self.job_queue.task_done()
-            self.result_queue.put(pickle.dumps(job, protocol=0))
-        return
 
 
 #-------------------------------------------------------------------------------
